Add Product Package functionality
Product Packages are a way for products that ship in there own packages to be defined properly for shipping estimates.  It is extremely useful for certain cases like a 200 lb product that ships in 5x 40lb boxes because 200lbs throws a UPS ship error.  This commit includes the ability to enter product packages through the admin interface.

The calculator now includes product packages in it's base calculation.  The user can now enter packages for a particular product and have those packages be used in the base freight calculation.  The calculator checks the products for over weight limit packages and will raise an error if the packages weigh too much.

Signed-off-by: Nathan Lowrie <nate@finelineautomation.com>

Fixes #55